wip: needs tests Add "health check address" concept to Host
This commit adds a "health check address" to each Host which is
logically separate from the destination address. It will be the same as
the destination address in all cases unless the host was specified using
EDS and the health_check_address field is set in the corresponding
Endpoint message.

This is a step toward supporting performing health checks against a
different address than traffic is routed to.

Signed-off-by: Michael Puncel <mpuncel@squareup.com>