xds protocol docs: clarify xDS server behavior when hints update
This addresses #3617. When a management server adds a cluster to CDS
that uses the EDS discovery type, there can be a race condition where
the EDS response is seen by Envoy before the CDS response, at which
point the endpoints for the new cluster are discarded (because Envoy
doesn't know about it yet). In this case it's expected that Envoy will
initiate a new EDS request with the new cluster name included as a
resource hint. This doc clarifies that the management server should send
a superfluous identical EDS response in this case so that Envoy can
learn the new endpoints.

Signed-off-by: Michael Puncel <mpuncel@squareup.com>